+++ ingeniamotion/fsoe.py
@@ -8,12 +8,12 @@ from fsoe_master.fsoe_master import (
     DictionaryItem,
     DictionaryItemInputOutput,
     MasterHandler,
-    StateData,
 )
 from ingenialink.enums.register import REG_ACCESS, REG_DTYPE
 from ingenialink.ethercat.register import EthercatRegister
 from ingenialink.pdo import RPDOMap, RPDOMapItem, TPDOMap, TPDOMapItem
 
+from ingeniamotion.enums import FSoEState
 from ingeniamotion.exceptions import IMTimeoutError
 from ingeniamotion.metaclass import DEFAULT_SERVO
 
@@ -138,7 +138,7 @@ class FSoEMasterHandler:
         state_reached = False
         init_time = time.time()
         while not state_reached:
-            state_reached = self.__master_handler.state == StateData
+            state_reached = self.state == FSoEState.DATA
             if timeout and (init_time + timeout) < time.time():
                 raise IMTimeoutError("The FSoE Master did not reach the Data state")
 
@@ -162,6 +162,11 @@ class FSoEMasterHandler:
     def safety_slave_pdu_map(self) -> TPDOMap:
         """The PDOMap used for the Safety Slave PDU."""
         return self.__safety_slave_pdu
+
+    @property
+    def state(self) -> FSoEState:
+        """Get the FSoE master state."""
+        return FSoEState(self.__master_handler.state.id)
 
 
 class PDUMapper:
@@ -358,6 +363,19 @@ class FSoEMaster:
         """
         master_handler = self.__handlers[servo]
         master_handler.wait_for_data_state(timeout)
+
+    def get_fsoe_master_state(self, servo: str = DEFAULT_SERVO) -> FSoEState:
+        """Get the servo's FSoE master handler state.
+
+        Args:
+            servo: servo alias to reference it. ``default`` by default.
+
+        Returns:
+            The servo's FSoE master handler state.
+
+        """
+        master_handler = self.__handlers[servo]
+        return master_handler.state
 
     def _delete_master_handler(self, servo: str = DEFAULT_SERVO) -> None:
         """Delete the master handler instance