+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ### Added
 - Method to check if drive is connected via a COM-KIT.
 - Property to check if the PDOs are active.
+- Method to get a servo's FSoE master handler state.
 
 ## [0.8.2] - 2024-07-15
 ### Added